Convert lint command tests to black-box package
